Fix compilation warning for taskpool

+++ Feldspar/C/taskpool.c
@@ -1,11 +1,11 @@
 //
 // Copyright (c) 2009-2011, ERICSSON AB
 // All rights reserved.
-// 
+//
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions are met:
-// 
-//     * Redistributions of source code must retain the above copyright notice, 
+//
+//     * Redistributions of source code must retain the above copyright notice,
 //       this list of conditions and the following disclaimer.
 //     * Redistributions in binary form must reproduce the above copyright
 //       notice, this list of conditions and the following disclaimer in the
@@ -13,10 +13,10 @@
 //     * Neither the name of the ERICSSON AB nor the names of its contributors
 //       may be used to endorse or promote products derived from this software
 //       without specific prior written permission.
-// 
+//
 // THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 // AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE 
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
 // DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
 // FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
 // DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
@@ -91,7 +91,7 @@ void spawn( void *closure )
 void *worker()
 {
     unsigned int self;
-    self = (unsigned int)pthread_self();
+    self = (unsigned long)pthread_self();
     log_1("worker %d - enter\n", self);
     struct taskpool *pool = feldspar_taskpool;
     void (*fun)();
@@ -130,7 +130,7 @@ void *worker()
             {
                 log_1("worker %d - sleep\n", self);
                 awake = 0;
-            } 
+            }
         }
         else
         {
